# [fix](scanner) Fix incorrect _max_thread_num in scanner context when many queries are running.