@zhiqiang-hhhh 2024-09-25
run buildall