@zhiqiang-hhhh 2024-10-09
run buildall