@zhiqiang-hhhh 2024-10-11
run buildall